+++ AngularApp/src/app/collapsible-menu/collapsible-menu.module.ts
@@ -2,10 +2,11 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { FormsModule } from '@angular/forms';
 import { CollapsibleMenuComponent } from './components/collapsible-menu/collapsible-menu.component';
-import { CollapsibleMenuItemComponent, SearchPipe, SearchMatchPipe } from './components/collapsible-menu-item/collapsible-menu-item.component';
+import { CollapsibleMenuItemComponent } from './components/collapsible-menu-item/collapsible-menu-item.component';
 import { SectionDividerComponent } from './components/section-divider/section-divider.component';
 import { SearchBoxComponent } from './components/search-box/search-box.component';
 import { MenuScrollComponent } from './components/menu-scroll/menu-scroll.component';
+import { SearchPipe, SearchMatchPipe } from './pipes/search.pipe';
 
 @NgModule({
   imports: [
@@ -13,6 +14,7 @@ import { MenuScrollComponent } from './components/menu-scroll/menu-scroll.compon
     FormsModule
   ],
   declarations: [CollapsibleMenuComponent, CollapsibleMenuItemComponent, SectionDividerComponent, SearchBoxComponent, MenuScrollComponent, SearchPipe, SearchMatchPipe],
-  exports: [CollapsibleMenuComponent, CollapsibleMenuItemComponent, SectionDividerComponent, SearchBoxComponent, MenuScrollComponent]
+  exports: [CollapsibleMenuComponent, CollapsibleMenuItemComponent, SectionDividerComponent, SearchBoxComponent, MenuScrollComponent, SearchPipe],
+  providers: [SearchPipe]
 })
 export class CollapsibleMenuModule { }
