+++ AngularApp/src/app/modules/main/main/main.component.ts
@@ -17,13 +17,6 @@ export class MainComponent implements OnInit {
   resourceName: string;
   resourceTypes: ResourceTypeState[] = [
     {
-      resourceType: null,
-      routeName: () => 'srid',
-      displayName: 'Support Request ID',
-      enabled: true,
-      caseId: true
-    },
-    {
       resourceType: ResourceType.Site,
       routeName: (name) => `sites/${name}`,
       displayName: 'App',
@@ -43,6 +36,13 @@ export class MainComponent implements OnInit {
       displayName: 'ARM Resource ID',
       enabled: true,
       caseId: false
+    },
+    {
+      resourceType: null,
+      routeName: () => 'srid',
+      displayName: 'Support Request ID',
+      enabled: true,
+      caseId: true
     }
   ];
